refactor: reusable component
