+++ packages/shared/src/components/post/ShareYouTubeContent.tsx
@@ -1,12 +1,11 @@
 import React, { ReactElement } from 'react';
-import Link from 'next/link';
 import PostSourceInfo from './PostSourceInfo';
 import { Post } from '../../graphql/posts';
 import { SharePostTitle } from './share';
 import { SharedLinkContainer } from './common/SharedLinkContainer';
 import { SharedPostLink } from './common/SharedPostLink';
 import YoutubeVideo from '../video/YoutubeVideo';
-import { formatReadTime } from '../utilities';
+import { formatReadTime, SelectableLink } from '../utilities';
 import { combinedClicks } from '../../lib/click';
 import ConditionalWrapper from '../ConditionalWrapper';
 
@@ -20,23 +19,6 @@ function ShareYouTubeContent({
   onReadArticle,
 }: ShareYouTubeContentProps): ReactElement {
   const isUnknownSource = post.sharedPost.source.id === 'unknown';
-  const onLinkClick = (e: React.MouseEvent) => {
-    const selection = globalThis?.window?.getSelection();
-    const hasSelection = selection?.anchorOffset !== selection?.focusOffset;
-
-    if (hasSelection) {
-      e.preventDefault();
-    }
-  };
-
-  const linkToPost = post.sharedPost.commentsPermalink;
-  const wrappedComponent = (wrapped) => (
-    <Link href={linkToPost}>
-      <a href={linkToPost} draggable="false" onClick={onLinkClick}>
-        {wrapped}
-      </a>
-    </Link>
-  );
 
   return (
     <>
@@ -47,7 +29,11 @@ function ShareYouTubeContent({
         wrapper={(node) => (
           <ConditionalWrapper
             condition={!isUnknownSource}
-            wrapper={wrappedComponent}
+            wrapper={(comp) => (
+              <SelectableLink href={post.sharedPost.commentsPermalink}>
+                {comp}
+              </SelectableLink>
+            )}
           >
             <SharedPostLink
               post={post}
