+++ packages/shared/src/components/utilities/index.ts
@@ -1,2 +1,3 @@
 export * from './common';
 export * from './Divider';
+export * from './SelectableLink';
